refactor(sync): restructure sync devices page with modular components and mixins
break down the large sync devices page into smaller, focused components
and introduce mixins for better code organization and maintainability
