+++ src/lib/presentation/ui/features/settings/components/sync_devices_tile.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:whph/presentation/ui/features/settings/components/settings_menu_tile.dart';
-import 'package:whph/presentation/ui/features/sync/pages/sync_devices_page.dart';
+import 'package:whph/presentation/ui/features/sync/pages/sync_devices_page/sync_devices_page.dart';
 import 'package:whph/presentation/ui/features/settings/constants/settings_translation_keys.dart';
 import 'package:whph/corePackages/acore/lib/utils/dialog_size.dart';
 import 'package:whph/presentation/ui/shared/services/abstraction/i_translation_service.dart';
